move request helper into helpers
